RAG_python_experiments/requirements.txt
langchain
openai
pinecone-client
tiktoken
pypdf
langchain-pinecone